Move contents of virtool.jobs_api to virtool.jobs
